pyproject.toml
[project]
name = "jiraah"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.1.3",
    "ipykernel>=6.29.5",
    "jira>=3.8.0",
    "pandas>=2.2.3",
    "pip>=24.3.1",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
]
